chore: fix declaration scope, temporary pointers, formats and license
Log: hoist the C struct forward declarations out of parameter scope so Swift receives typed pointers, pass local values instead of pointers that outlive their closure or NSString, cast the PRIu64 diagnostics, refresh LICENSE-iperf to the 3.21 text, and document that .running carries no server-ready signal.

+++ Sources/IperfCLib/iperf_api.c
@@ -3898,7 +3898,7 @@ iperf_print_intermediate(struct iperf_test *test)
 	    iperf_time_diff(&irp->interval_start_time, &irp->interval_end_time, &temp_time);
 	    double interval_len = iperf_time_in_secs(&temp_time);
 	    if (test->debug) {
-		printf("interval_len %f bytes_transferred %" PRIu64 "\n", interval_len, irp->bytes_transferred);
+		printf("interval_len %f bytes_transferred %" PRIu64 "\n", interval_len, (uint64_t) irp->bytes_transferred);
 	    }
 
 	    /*
